# backend/openapi/api-design.yaml
openapi: 3.1.0
info:
  title: syumatsucamera API Design Draft
  version: 0.1.0
  description: |
    本ファイルは実装前のAPI設計をGit管理するためのドラフトです。
    `AGENTS.md` と各種要件mdを元にした初版であり、実装完了後の自動生成schemaではありません。
servers:
  - url: https://syumatsucamera.com
    description: Public API
  - url: https://cms.syumatsucamera.com
    description: CMS API
...
security: []
...
paths:
...
  /api/users/session/me:
    get:
      tags:
        - users
      summary: CMS利用中のユーザー情報を返す
      description: |
        認証: CloudflareAccessJwt
        権限: admin または author
        permission_class: AuthorAdminReadWrite
        lock_token を検証後、必要な場合のみ tmp/{lock_token} ディレクトリを作成する
      operationId: getCmsSessionMe
      security:
        - CloudflareAccessJwt:
            - admin
        - CloudflareAccessJwt:
            - author
      x-permission-class: AuthorAdminReadWrite
      responses:
        "200":
          description: 正常
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yaml#/components/schemas/UsersSessionUser"
        "401":
          description: 認証失敗
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yaml#/components/schemas/CommonDetailError"
...
  /api/users/{user_id}/invite:
    post:
      tags:
        - users
      summary: 管理者が仮登録ユーザーの招待URLを発行する
      description: |
        認証: CloudflareAccessJwt
        権限: admin
      operationId: issueUserInvitation
      security:
        - CloudflareAccessJwt:
            - admin
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: 発行成功
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yaml#/components/schemas/UsersActivationIssueResponse"
        "404":
          description: ユーザーが存在しない
          content:
            application/json:
              schema:
                $ref: "./components/schemas.yaml#/components/schemas/CommonDetailError"
...
components:
  securitySchemes:
    CloudflareAccessJwt:
      type: apiKey
      in: header
      name: Cf-Access-Jwt-Assertion
      description: 記事管理 / ユーザー管理 / OGP管理 / 招待本登録などの Access 保護対象APIで Cloudflare Access が付与するJWT
  parameters:
...
    UserId:
      name: user_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...